+++ frontend/www/lib/angularjs-slider/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 2.8.0 (2016-02-08)
+## Features
+- Add a `getPointerColor` option to dynamically change the pointers color (#253).
+
+# 2.7.1 (2016-02-06)
+## Fix
+- Fix high label positioning when size is different than the ceil one.
+
+# 2.7.0 (2016-02-06)
+## Features
+- Add an `enforceStep` option (defaults to true) (#246).
+- Add a `showSelectionBarFromValue` options (#250).
+- Use jqLite html() method to display label values so the translate function can return formated content (#251).
+- Pass a label string as third arg to the `translate` function to differentiate the labels (#252).
+
+## Fix
+- Improve combined label position and show only one value if min==max (#245).
+
 # 2.6.0 (2016-01-31)
 ## Features
 - Add a `noSwitching` option to prevent the user from switching the min and max handles (#233).